refactor(main|extract): centralize & normalize normalizations
